+++ ouds_theme_contract/lib/theme/tokens/components/ouds_components_tokens.dart
@@ -11,6 +11,7 @@
  */
 
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_badge_tokens.dart';
+import 'package:ouds_theme_contract/theme/tokens/components/ouds_bar_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_buttonMono_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_button_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_checkbox_tokens.dart';
@@ -26,8 +27,6 @@ import 'package:ouds_theme_contract/theme/tokens/components/ouds_skeleton_tokens
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_switch_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_tag_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_textInput_tokens.dart';
-
-import 'ouds_linkMono_tokens.dart';
 
 class OudsComponentsTokens {
   final OudsButtonTokens button;
@@ -46,6 +45,7 @@ class OudsComponentsTokens {
   final OudsPinCodeInputTokens pinCodeInput;
   final OudsLinkTokens link;
   final OudsLinkMonoTokens linkMono;
+  final OudsBarTokens bar;
 
   const OudsComponentsTokens({
     required this.button,
@@ -64,5 +64,6 @@ class OudsComponentsTokens {
     required this.pinCodeInput,
     required this.link,
     required this.linkMono,
+    required this.bar,
   });
 }